requirements.txt
pydantic
pillow
numpy
